Query up to 10 first pages for workflow runs
GitHub api returns 30 results per page; we raise the limit to 100 (max
allowed) and limit ourselves to max 10 pages.

We're querying for master branch, so in the usual case results will be
found on the first page anyway. In the worst case, we will do 33 GitHub
API calls (30x workflows/runs + 3x artifacts) and retrieve info about
last 1000 builds.

We could implement rate limiting to ensure the whole build history is
traversed, but it's rather pointless as GitHub might purge build
artifacts after ~30 days anyway.

This solution is not perfect, but it's "good enough" - at least as long
as we're doing regular merges to master branch :)