Update the vision for 2022
One of Staging's stated goals is to encourage new contributors by
removing barriers to entry.

One of these barriers was in fact Staging's constrained
gaming-specific wording. To keep with our goal of growing the
team, this commit puts general purpose and gaming-focussed
emulation improvements on equal footing with a new caveat being
that the changes (in either case) need to be valuable.

Previously we were using gaming as a substitute for "value",
meaning that if one or more games use a given emulation feature,
then that feature could be deemed valuable. This new change means
we can assess the value or merit of filling an emulation hole
given the broader context of the value it brings to users.

This commit also tightens up the wording in a couple of other
areas (but keeps the same meaning and intent).