Add Wengier to the 0.79 maintainers
As the distinction between roles was not documented or explained, that is:

maintainers:

   Are focused on the internal health and maintainability of the project. They
   are the primary PR reviewers, dependency managers, and build tool updaters.
   They are concerned with non-user-facing aspects of keeping the project
   running smooth for authors.

authors:

   They bring meaningful features, fixes, and documentation forward for the
   user. Their contributions progress the baseline of the project forward.

The maintainer helps keep the stage set upon which the authors work is
performed.

The two roles are equally critical to keeping a project both health *and* moving
forward in a meaningful way.

Without authors, the project lapses into "maintenance-mode", where it's kept
alive for the sake of it, but without meaningful progress.

---

As these were not explained during the 0.79 period, I suggest we stick to the
previously less distinct (or assumed) roles.

For 0.80, these roles will be documented in the wiki so team members can
decide (if they care) to put effort into one or the other.