Add initial 0.80 release notes
Content authors:
 - FeralChild64, for mouse and more section content
 - JohnNovak, for Ability to use local configs via Finder on macOS section content.
 - kcgen, for balance of the document.

Thanks to JohnNovak for the template refinement suggestion and detailed review feedback.

--- Template suggestion: ---

May I suggest the following structure going forward:

- Summary

- How to upgrade

- New features
  - ReelMagic support
  - Mouse improvements
    - Dedicated configuration section
    - Improved support for 3rd-party Windows mouse drivers
    - Mouse-related help in the title bar
  - Ability to use local configs via Finder on macOS

- General improvements
  - more command
  - On-the-fly changes
  - Post-rendered screenshots
  - More bundled OpenGL shaders
  - Further additions and improvements
        <only list the rest here>

- Fixes and regressions

- Translation improvements

    <I think translations should be a separate
    section, and there was only a single README
    update, that's developer doco and shouldn't be
    mentioned>

- Contributors

- Thank you

---

most notably:

- this is just a slight refinement of the format of
the last two release notes — there's no need to
change it up in every release

- we should use hierarchical headings (maybe not so
important right now, but that will show up nicely on
the right-side "Table of contents" sidebar in the new
design which will be launched hopefully this year!)

- i strongly suggest to drop "Title Case" and just
use "Normal sentence case" for headings, which is in
line with current best-practices

So basically the 1st level headings will be fixed,
and will never change. of course, if there's zero
fixes or regressions in the release, then we can omit
that heading... btw, i've converted the existing
release notes to more or less adhere to this format
(i tried to strike a balance between massaging them
into this format and a complete rewrite)

---